requirements.txt
discord.py
motor
config